requirements.txt
APScheduler==3.6.0
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
decorator==4.4.1
dj-database-url==0.5.0
Django==2.2.2
django-background-tasks==1.2.0
django-compat==1.0.15
django-cors-headers==3.2.0
django-csp==3.5
django-heroku==0.3.1
django-redis-cache==2.1.0
djangorestframework==3.9.4
fsspec==0.6.0
gcsfs==0.4.0
google-api-core==1.14.3
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-cloud==0.34.0
google-cloud-core==1.0.3
google-cloud-pubsub==1.0.2
google-cloud-storage==1.23.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
grpc-google-iam-v1==0.12.3
grpcio==1.25.0
gunicorn==19.9.0
idna==2.8
joblib==0.14.0
Markdown==3.1.1
numpy==1.16.4
oauthlib==3.1.0
pandas==0.24.2
plotly==4.3.0
protobuf==3.10.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
pyasn1==0.4.8
pyasn1-modules==0.2.7
Pygments==2.4.2
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
redis==3.3.11
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
scikit-learn==0.21.3
scipy==1.3.2
six==1.12.0
sqlparse==0.3.0
tzlocal==1.5.1
urllib3==1.25.3
whitenoise==4.1.2
